Fixing many of the issues with new tsc compile method

+++ src/twitch/twitch.service.ts
@@ -3,15 +3,20 @@
  */
 import { ClientCredentialsAuthProvider, RefreshingAuthProvider, exchangeCode, AccessToken, RefreshConfig } from '@twurple/auth';
 import { ApiClient, HelixPrivilegedUser } from '@twurple/api';
-import { EventSubListener } from '@twurple/eventsub';
+import { EventSubListener, EventSubSubscription } from '@twurple/eventsub';
 import { NgrokAdapter } from '@twurple/eventsub-ngrok';
 
 /** 
  * Necessary Defines
  */
-const twitchClientAuth = {
-    clientId: process.env.TWITCH_CLIENT_ID,
-    clientSecret: process.env.TWITCH_CLIENT_SECRET
+var twitchClientAuth: {clientId: string, clientSecret: string};
+if(typeof process.env.TWITCH_CLIENT_ID == 'string' && typeof process.env.TWITCH_CLIENT_SECRET == 'string') {
+    twitchClientAuth = {
+        clientId: process.env.TWITCH_CLIENT_ID,
+        clientSecret: process.env.TWITCH_CLIENT_SECRET
+    }
+} else {
+    process.exit();
 }
 // const clientId = process.env.TWITCH_CLIENT_ID;
 // const clientSecret = process.env.TWITCH_CLIENT_SECRET;
@@ -19,8 +24,15 @@ const authProvider = new ClientCredentialsAuthProvider(twitchClientAuth.clientId
 const apiClient = new ApiClient({ authProvider });
 // const refreshingAuthProvider
 
+var twitch_callback: string;
+if(typeof process.env.TWITCH_CALLBACK == 'string') {
+    twitch_callback = process.env.TWITCH_CALLBACK;
+} else {
+    process.exit(1);
+}
+
 /**
- * Create Listender
+ * Create Listener
  */
 
 const listenerSecret = get_unique_reference_number();
@@ -51,9 +63,11 @@ export const init_listener = async () => {
 
 export const auth_twitch = async (twitch_code: string): Promise<AccessToken | false> => {
     try {
-        const response = await exchangeCode(twitchClientAuth.clientId, twitchClientAuth.clientSecret, twitch_code, process.env.TWITCH_CALLBACK);
-        console.table(response);
-        return response;
+        if(typeof twitchClientAuth.clientId == 'string' && typeof twitchClientAuth.clientSecret == 'string') {
+            const response = await exchangeCode(twitchClientAuth.clientId, twitchClientAuth.clientSecret, twitch_code, twitch_callback);
+            console.table(response);
+            return response;
+        }
     }
     catch {
         console.log("Error with exchanging code for access token");
@@ -70,16 +84,16 @@ export const auth_twitch = async (twitch_code: string): Promise<AccessToken | fa
 export class Twitch_Streamer {
     twitch_service = true;
     private _unique_id: string;
-    private _twitch_id: string = null;
-    private _name: string = null;
-    private _access_token: AccessToken;
-    private _refresh_config: RefreshConfig;
-    private _refreshingAuthProvider: RefreshingAuthProvider;
-    private _apiClient: ApiClient;
+    private _twitch_id: string | null = null;
+    private _name: string | null = null;
+    private _access_token: AccessToken | null = null;
+    private _refresh_config: RefreshConfig | null = null;
+    private _refreshingAuthProvider: RefreshingAuthProvider | null = null;
+    private _apiClient: ApiClient | null = null;
     // private _twitchListener: EventSubListener;
-    private _onlineSubscription;
-    private _offlineSubscription;
-    private _store_data_callback;
+    private _onlineSubscription: EventSubSubscription | null = null;
+    private _offlineSubscription: EventSubSubscription | null = null;
+    private _store_data_callback: Function | null = null;
 
     constructor(unique_id: string, options: {twitch_id?: string, accessToken?: AccessToken, listenerSecret?: string}) {
         this._unique_id = unique_id;
@@ -91,19 +105,19 @@ export class Twitch_Streamer {
         }
     }
 
-    public set twitch_id(v : string) {
+    public set twitch_id(v : string | null) {
         this._twitch_id = v;
     }
     
-    public get twitch_id() : string {
+    public get twitch_id() : string | null {
         return this._twitch_id;
     }
     
-    public set name(v : string) {
+    public set name(v : string | null) {
         this._name = v;
     }
 
-    public get name() : string {
+    public get name() : string | null {
         return this._name;
     }
 
@@ -111,38 +125,40 @@ export class Twitch_Streamer {
         this._access_token = v;
     }
 
-    public async retreive_twitch_data(store_data): Promise<HelixPrivilegedUser | false> {
+    public async retreive_twitch_data(store_data: Function): Promise<HelixPrivilegedUser | false> {
         this._store_data_callback = store_data;
         this._refresh_config = {
             clientId: twitchClientAuth.clientId,
             clientSecret: twitchClientAuth.clientSecret,
             onRefresh: this.store_twitch_access_token
         }
-        this._refreshingAuthProvider = new RefreshingAuthProvider(this._refresh_config, this._access_token);
-        this._apiClient = new ApiClient({ authProvider: this._refreshingAuthProvider }); //TODO: Can set logger details with this
+        if(typeof this._access_token == 'string') {
+            this._refreshingAuthProvider = new RefreshingAuthProvider(this._refresh_config, this._access_token);
+            this._apiClient = new ApiClient({ authProvider: this._refreshingAuthProvider }); //TODO: Can set logger details with this
 
-        try {
-            await this.deleteSubscriptions();
-            console.log("Deleting Subscriptions");
+            try {
+                await this.deleteSubscriptions();
+                console.log("Deleting Subscriptions");
 
-            const data: HelixPrivilegedUser = await this._apiClient.users.getMe();
+                const data: HelixPrivilegedUser = await this._apiClient.users.getMe();
 
-            console.log(data);
+                console.log(data);
 
-            console.log("Retreived me");
+                console.log("Retreived me");
 
-            if(!this.twitch_id) {
-                console.log("Retreiving this user's data from twitch for first time");
-                this.twitch_id = data.id;
-                this.name = data.name;
+                if(!this.twitch_id) {
+                    console.log("Retreiving this user's data from twitch for first time");
+                    this.twitch_id = data.id;
+                    this.name = data.name;
+                }
+
+                if(data) {
+                    return data;
+                }
+            } catch (e){
+                console.log(`Failed to retrieve user data: ${e}`);
+                return false;
             }
-
-            if(data) {
-                return data;
-            }
-        } catch (e){
-            console.log(`Failed to retrieve user data: ${e}`);
-            return false;
         }
         console.log("Data not received for user");
         return false;
@@ -150,7 +166,9 @@ export class Twitch_Streamer {
 
     public async store_twitch_access_token() {
         console.log("Storing Twitch Access Token");
-        this._store_data_callback(this._unique_id, this._access_token);
+        if(this._store_data_callback !== null) {
+            this._store_data_callback(this._unique_id, this._access_token);
+        }
     }
 
     public async setup_live_subscriptions(state_functions: Array<any>) {
@@ -162,81 +180,51 @@ export class Twitch_Streamer {
         // }
         console.log(`Setting up live subscriptions for ${this.name}`);
         // setTimeout(async () => {
-        try {
-            console.log(`Got to online events: ${this.twitch_id}`);
+        if(typeof this.twitch_id == 'string') {
+            try {
+                console.log(`Got to online events: ${this.twitch_id}`);
 
-            this._onlineSubscription = await twitch_listener.subscribeToStreamOnlineEvents(this.twitch_id, e => {
-                console.log(`${e.broadcasterDisplayName} just went live!`);
-                state_functions[0](this.twitch_id);
-            });
-        } catch (error) {
-            console.log(`Error: ${error}`);
+                this._onlineSubscription = await twitch_listener.subscribeToStreamOnlineEvents(this.twitch_id, e => {
+                    console.log(`${e.broadcasterDisplayName} just went live!`);
+                    state_functions[0](this.twitch_id);
+                });
+            } catch (error) {
+                console.log(`Error: ${error}`);
+            }
+                // setTimeout(async () => {
+            try {
+                console.log(`Got to offline events: ${this.twitch_id}`);
+                this._offlineSubscription = await twitch_listener.subscribeToStreamOfflineEvents(this.twitch_id, e => {
+                    console.log(`${e.broadcasterDisplayName} just went offline.`);
+                    state_functions[1](this.twitch_id);
+                });
+            } catch (error) {
+                console.log(`Error: ${error}`);
+            }
         }
-            // setTimeout(async () => {
-        try {
-            console.log(`Got to offline events: ${this.twitch_id}`);
-            this._offlineSubscription = await twitch_listener.subscribeToStreamOfflineEvents(this.twitch_id, e => {
-                console.log(`${e.broadcasterDisplayName} just went offline.`);
-                state_functions[1](this.twitch_id);
-            });
-        } catch (error) {
-            console.log(`Error: ${error}`);
-        }
-
             // }, 1000);
         // }, 1000);
 
     }
 
     public async cancel_live_subscriptions() {
-        try {
-            await this._onlineSubscription.stop();
+        if(this._offlineSubscription !== null && this._onlineSubscription !== null) {
+            try {
+                await this._onlineSubscription.stop();
 
-            await this._offlineSubscription.stop();
-        } catch (error) {
-            console.log(error);
+                await this._offlineSubscription.stop();
+            } catch (error) {
+                console.log(error);
+            }
         }
     }
 
     public async deleteSubscriptions() {
-        await this._apiClient.eventSub.deleteAllSubscriptions();
+        if(this._apiClient !== null) {
+            await this._apiClient.eventSub.deleteAllSubscriptions();
+        }
     }
 }
-
-// export const deleteAllSubscriptions = async () => {
-//     // var subscriptions_data = await apiClient.eventSub.getSubscriptions();
-//     // console.table(subscriptions_data);
-//     await apiClient.eventSub.deleteAllSubscriptions();
-// }
-
-/**
- * Twitch API Functions
- */
-
-// Will need to be in the twitch service
-// Setting up auth for twitch
-
-// async function create_listener() {
-//     const listenerSecret = get_unique_reference_number();
-
-//     const listener = new EventSubListener({
-//         apiClient,
-//         adapter: new NgrokAdapter(),
-//         secret: listenerSecret
-//     });
-//     await listener.listen();
-// }
-
-// async function setup_live_subscriptions(twitch_name: string) {
-//     const onlineSubscription = await listener.subscribeToStreamOnlineEvents(userId, e => {
-//         console.log(`${e.broadcasterDisplayName} just went live!`);
-//     });
-
-    
-//     const offlineSubscription = await listener.subscribeToStreamOfflineEvents(userId, e => {
-//         console.log(`${e.broadcasterDisplayName} just went offline.`);
-//     });
-// }
 
 /**
  * Service Functions
